+++ src/main/java/org/scion/jpan/internal/ControlServiceGrpc.java
@@ -17,6 +17,7 @@ package org.scion.jpan.internal;
 import io.grpc.*;
 import io.grpc.okhttp.OkHttpChannelBuilder;
 import java.util.concurrent.TimeUnit;
+
 import org.scion.jpan.ScionRuntimeException;
 import org.scion.jpan.ScionUtil;
 import org.scion.jpan.proto.control_plane.Seg;
@@ -31,6 +32,7 @@ public class ControlServiceGrpc {
   private final LocalTopology localAS;
   private ManagedChannel channel;
   private SegmentLookupServiceGrpc.SegmentLookupServiceBlockingStub grpcStub;
+  private final int deadLineMs;
 
   public static ControlServiceGrpc create(LocalTopology localAS) {
     return new ControlServiceGrpc(localAS);
@@ -38,6 +40,7 @@ public class ControlServiceGrpc {
 
   private ControlServiceGrpc(LocalTopology localAS) {
     this.localAS = localAS;
+    this.deadLineMs = Config.getControlPlaneTimeoutMs();
   }
 
   public void close() {
@@ -69,12 +72,19 @@ public class ControlServiceGrpc {
     grpcStub = SegmentLookupServiceGrpc.newBlockingStub(channel);
   }
 
+  private SegmentLookupServiceGrpc.SegmentLookupServiceBlockingStub getStub() {
+    // This is a deadline, not a timeout. It counts from the time the "with..." is called.
+    // See also https://github.com/grpc/grpc-java/issues/1495
+    // and https://github.com/grpc/grpc-java/issues/4305#issuecomment-378770067
+    return grpcStub.withDeadlineAfter(deadLineMs, TimeUnit.MILLISECONDS);
+  }
+
   public synchronized Seg.SegmentsResponse segments(Seg.SegmentsRequest request) {
     if (channel == null) {
       return segmentsTryAll(request);
     }
     try {
-      return grpcStub.segments(request);
+      return getStub().segments(request);
     } catch (StatusRuntimeException e) {
       if (e.getStatus().getCode().equals(Status.Code.UNAVAILABLE)) {
         closeChannel();
@@ -107,7 +117,7 @@ public class ControlServiceGrpc {
     for (LocalTopology.ServiceNode node : localAS.getControlServices()) {
       initChannel(node.ipString);
       try {
-        return grpcStub.segments(request);
+        return getStub().segments(request);
       } catch (StatusRuntimeException e) {
         if (e.getStatus().getCode().equals(Status.Code.UNAVAILABLE)) {
           LOG.warn("Error connecting to control service: {}", node.ipString);
@@ -115,7 +125,7 @@ public class ControlServiceGrpc {
           continue;
         }
         // Rethrow the exception if it's not UNAVAILABLE
-        throw e;
+        throw new ScionRuntimeException("Error getting segments from control service.", e);
       }
     }
     throw new ScionRuntimeException(